# internal/ent/schema/user.go
package schema

import (
	"github.com/facebook/ent/dialect"

	"github.com/facebook/ent"
	"github.com/facebook/ent/schema/field"
)

// User holds the schema definition for the User entity.
type User struct {
	ent.Schema
}

// Mixin of the User.
func (User) Mixin() []ent.Mixin {
	return []ent.Mixin{}
}

func (User) Config() ent.Config {
	return ent.Config{
		Table: "user",
	}
}

// Fields of the User.
func (User) Fields() []ent.Field {
	return []ent.Field{
		field.Int64("id").Comment(`主键`),
		field.String("identity").Optional().Comment(`标识`),
		field.Int32("user").Default(0).Comment(`类型：XXX`),
		field.Int32("is_disable").Default(0).Comment(`是否禁用：1-禁用，0-正常，默认：0`),
		field.Time("created_at").Optional().SchemaType(map[string]string{dialect.MySQL: "datetime"}).Comment(`创建时间`),
		field.Time("updated_at").Optional().SchemaType(map[string]string{dialect.MySQL: "datetime"}).Comment(`更新时间`),
		field.Time("deleted_at").Optional().SchemaType(map[string]string{dialect.MySQL: "datetime"}).Comment(`删除时间`),
	}
}

// Edges of the User.
func (User) Edges() []ent.Edge {
	return nil
}